Move helper to utils.py
